src/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-filebrowser-no-grappelli==3.8.0
django-tinymce4-lite==1.8.0
jsmin==2.2.2
Pillow==7.2.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
